hw_18/requirements.txt
yoyo-migrations==8.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
